+++ Application/Models/Тreatments/ListТreatmentModel.cs
@@ -19,7 +19,7 @@ namespace Application.Models.Тreatments
 
         public string TreatmentType { get; init; } = default!;
 
-        public int AmountOfFuel { get; init; }
+        public decimal AmountOfFuel { get; init; }
 
         public decimal FuelPrice { get; init; }
 
@@ -27,7 +27,7 @@ namespace Application.Models.Тreatments
 
         public string ArticleName { get; set; } = default!;
 
-        public int ArticleQuantity { get; init; }
+        public decimal ArticleQuantity { get; init; }
 
         public decimal ArticlePrice { get; init; }
