+++ WebUI/ServicesModel/Тreatment/ТreatmentBaseModel.cs
@@ -6,13 +6,13 @@
 
         public DateTime Date { get; set; } = DateTime.Now;
 
-        public int AmountOfFuel { get; set; }
+        public decimal AmountOfFuel { get; set; }
 
         public decimal FuelPrice { get; set; }
 
         public int ArticleId { get; set; }
 
-        public int ArticleQuantity { get; set; }
+        public decimal ArticleQuantity { get; set; }
 
         public decimal ArticlePrice { get; set; }
     }
